Move the compile-bench job impl to its own module.
